requirements.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
Appium-Python-Client==0.48
atomicwrites==1.3.0
attrs==19.3.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
chromedriver-autoinstaller==0.2.0
colorama==0.4.1
google-api-core==1.19.0
google-api-python-client==1.9.1
google-auth==1.16.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.9
importlib-metadata==0.23
more-itertools==7.2.0
oauthlib==3.1.0
packaging==19.2
pluggy==0.13.0
protobuf==3.12.2
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.2
pytest==5.2.2
pytest-ordering==0.6
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
selenium==3.141.0
six==1.12.0
uritemplate==3.0.1
urllib3==1.25.6
wcwidth==0.1.7
zipp==0.6.0
